# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Twindix Admin API",
        "description": "API documentation for Twindix Admin Panel - Access Control System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/regions": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "List all regions",
                "operationId": "55a86a3e6cad996799d0f7eb926559e1",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Regions"
                ],
                "summary": "Create new region",
                "operationId": "a774206732b4fd23848a53a1eb143ada",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/regions/{id}": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Get single region",
                "operationId": "5f7d3aae22684e0542de09960daa956f",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Regions"
                ],
                "summary": "Update region",
                "operationId": "b12106b8a7dd961d198f34d8c3055e40",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Regions"
                ],
                "summary": "Delete region",
                "operationId": "a88665c37bfa34ad8df0d7b54eda79d1",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/countries": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "List all countries",
                "operationId": "be763e06520c2693aa4fafdfce356bb2",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/super-admins": {
            "get": {
                "tags": [
                    "Super Admins"
                ],
                "summary": "List super admins",
                "operationId": "e278886ff195dc52eb233883e797b83f",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Super Admins"
                ],
                "summary": "Create super admin",
                "operationId": "29730cca90859b23d5b04d803b3fd7c8",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/super-admins/{id}": {
            "get": {
                "tags": [
                    "Super Admins"
                ],
                "summary": "Get super admin",
                "operationId": "0712f05846b020de43e512019197de87",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Super Admins"
                ],
                "summary": "Update super admin",
                "operationId": "6b8cceff44f974c889239d7d0721e699",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Super Admins"
                ],
                "summary": "Delete super admin",
                "operationId": "012d4b5df8ff836e8e7443a2d77cdfe7",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/super-admins/{id}/pause": {
            "post": {
                "tags": [
                    "Super Admins"
                ],
                "summary": "Pause account",
                "operationId": "3e5014153e102559d15e95d59c9330e1",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/super-admins/{id}/resume": {
            "post": {
                "tags": [
                    "Super Admins"
                ],
                "summary": "Resume account",
                "operationId": "056e94025d7dc446ebb048216c0ce2f0",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admins": {
            "post": {
                "tags": [
                    "Super Admins"
                ],
                "summary": "Create admin under super admin",
                "operationId": "ca8916473aae3eb99006e1d95923d112",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}